batch_layer/requirements.txt
kafka-python==2.3.0
pandas==2.3.3
PyYAML==6.0.3